Parse diskutil stdout using it's plist output, rather than regex
The current method of using regex to parse diskutil's stdout is somewhat
brittle (plus I'm not confident in my regex-foo). This commit uses MacOS's
plutil to convert diskutil's plist output to JSON, and then parses the JSON to
the resulting structs.

I'd rather parse the plist output in Go directly, but it seems that this is not
trivial with the stdlib [0], and I'd rather not add dependencies for this small
project.

[0]: https://stackoverflow.com/questions/38657880/parsing-plist-xml