Specify source and target volumes with -source and -target flags
This commit also allows multiple targets to be specified, with each being
cloned successively. As far as I can tell, it's not possible for `asr` to
restore to multiple target volumes at once. It's also not possible to call
`asr` multiple times at once for the same source volume, as the volume is
unmounted when the restore begins.